+++ packages/logo/src/components/glyphs/Glyph.tsx
