+++ packages/logo/src/components/logos/GlyphStack.tsx
