+++ packages/logo/src/components/logos/RNALogo.tsx
@@ -1,20 +1,22 @@
-import { A, C, G, T } from "../glyphs";
-import { Logo } from "./logo";
-import { Alphabets, DNALogoProps } from "./types";
+import React from 'react';
+
+import { A, C, G, U } from '../glyphs';
+import { Logo } from './Logo';
+import { Alphabets, RNALogoProps } from './types';
 
 /**
- * Represents the DNA alphabet, with the four nucleotides colored
+ * Represents the RNA alphabet, with the four nucleotides colored
  * in a similar scheme to the MEME default.
  */
-export const DNAAlphabet: Alphabets[] = [
-  { component: [A], regex: ["A"], color: ["red"] },
-  { component: [C], regex: ["C"], color: ["blue"] },
-  { component: [G], regex: ["G"], color: ["orange"] },
-  { component: [T], regex: ["T"], color: ["#228b22"] },
+export const RNAAlphabet: Alphabets[] = [
+  { component: [A], regex: ['A'], color: ['red'] },
+  { component: [C], regex: ['C'], color: ['blue'] },
+  { component: [G], regex: ['G'], color: ['orange'] },
+  { component: [U], regex: ['U'], color: ['seagreen'] },
 ];
 
 /**
- * Renders a logo with the DNA alphabet, with nucleotides colored similarly to the MEME default.
+ * Renders a logo with the RNA alphabet, with nucleotides colored similarly to the MEME default.
  *
  * @prop ppm position probability matrix. Rows are positions and should sum to 1.0; columns are nucleotides,
  *           alphabetically. If this is provided, it takes precedence over PFM in computing symbol heights.
@@ -23,6 +25,6 @@ export const DNAAlphabet: Alphabets[] = [
  * @prop startPos number to assign the first position in the logo; defaults to 1.
  * @prop yAxisMax if set, uses an explicit maximum value for the y-axis rather than the total number of bits possible. This is ignored in FREQUENCY mode.
  */
-export function DNALogo(props: DNALogoProps) {
-  return <Logo {...props} alphabet={DNAAlphabet} />;
-}
+export const RNALogo: React.FC<RNALogoProps> = props => (
+  <Logo {...props} alphabet={RNAAlphabet} />
+);
