+++ packages/logo/src/components/logos/XAxis.tsx
