+++ packages/logo/src/components/logos/YAxis.tsx
