+++ packages/logo/src/components/logos/YAxisFrequency.tsx
