+++ packages/logo/src/components/logos/YAxisWithNegatives.tsx
